Revert "Merge pull request #131 from COS301-SE-2023/rework"
This reverts commit ff4aacc1f1915f97f8dacbd332aa347503ad777e, reversing
changes made to 2b4d6ceddd6d8e6f1c39a29f19e5907491393de0.

+++ src/libs/shared-utils/models/ApplianceModel.cs
@@ -5,7 +5,7 @@ public class ApplianceModel
     public int powerUsage { get; set; }
     public bool? recordState { get; set; }
     public int quantity { get; set; } = 0;
-    public double durationUsed { get; set; }
+    public float durationUsed { get; set; }
     public string? fade { get; set; } = "fade-in-enter-active";
     public string? name { get; set; } = "Default";
 }
